uniformized requests - requests import is in external import group - workspace id is always required in deployment or as argument - Token is found using click find_object - request exception is always caught and exits cleanly

+++ Babylon/groups/powerbi/groups/workspace/commands/delete.py
@@ -2,6 +2,7 @@ import logging
 from typing import Optional
 
 import requests
+from azure.core.credentials import AccessToken
 from click import Context
 from click import command
 from click import option
@@ -17,7 +18,7 @@ logger = logging.getLogger("Babylon")
 
 @command()
 @pass_context
-@option("-w", "--workspace", "override_workspace_id", required=False, type=QueryType())
+@option("-w", "--workspace", "override_workspace_id", help="PowerBI workspace ID", type=QueryType())
 @require_deployment_key("powerbi_workspace_id", required=False)
 @option(
     "-f",
@@ -26,12 +27,16 @@ logger = logging.getLogger("Babylon")
     is_flag=True,
     help="Don't ask for validation before delete",
 )
-def delete(ctx: Context, powerbi_workspace_id: str, override_workspace_id: Optional[str], force_validation: bool):
+def delete(ctx: Context, powerbi_workspace_id: str, override_workspace_id: Optional[str],
+           force_validation: bool) -> CommandResponse:
     """Delete WORKSPACE_NAME from Power Bi APP"""
-    access_token = ctx.parent.obj.token
+    access_token = ctx.find_object(AccessToken).token
     workspace_id = override_workspace_id or powerbi_workspace_id
+    if not workspace_id:
+        logger.error("A workspace id is required either in your config or with parameter '-w'")
+        return CommandResponse(status_code=CommandResponse.STATUS_ERROR)
     if not force_validation and not confirm_deletion("Power Bi Workspace", workspace_id):
-        return
+        return CommandResponse(status_code=CommandResponse.STATUS_ERROR)
     url_delete = f'https://api.powerbi.com/v1.0/myorg/groups/{workspace_id}'
     header = {'Content-Type': 'application/json', 'Authorization': f'Bearer {access_token}'}
     response = requests.delete(url=url_delete, headers=header)
